Classifier/requirements.txt
matplotlib==3.0.3
numpy==1.16.2
torchvision==0.2.2.post3
torch==1.1.0
Pillow==6.1.0
tensorboardX==1.8
